requirements.txt
requests
pytz
psycopg2-binary